requirements.txt
PySide6
QScintilla
furo
seabreeze
pyqtgraph
silx
numpy
scipy
pandas
lmfit
sympy
scikit-learn
statsmodels
matplotlib
seaborn
openpyxl
xlsxwriter
pillow
h5py
requests
pyyaml
pyserial
pyvisa

